fix tests and repackage into module (lexer & bool_array)
