+++ src/bool_array.c
@@ -1,4 +1,3 @@
-#include "minishell.h"
 #include <stdbool.h>
 #include "libft.h"
