+++ src/lexer/lexer_support.c
@@ -1,0 +1,36 @@
+#include "struct.h"
+
+int	str_cchr(const char *s, char c)
+{
+	int	i;
+	int	count;
+
+	i = 0;
+	count = 0;
+	if (!s)
+		return (0);
+	while (s[i])
+	{
+		if (s[i] == c)
+			count++;
+		i++;
+	}
+	return (count);
+}
+
+void	count_number(char *s, struct s_lexer *input)
+{
+	*input = (struct s_lexer){
+		.singlequotes = str_cchr(s, '\''),
+		.doublequotes = str_cchr(s, '"'),
+		.open_curly_brackets = str_cchr(s, '{'),
+		.close_curly_brackets = str_cchr(s, '}'),
+		.open_square_brackets = str_cchr(s, '['),
+		.close_square_brackets = str_cchr(s, ']'),
+		.open_parentheses = str_cchr(s, '('),
+		.close_parentheses = str_cchr(s, ')'),
+		.redir_greater = str_cchr(s, '>'),
+		.redir_smaller = str_cchr(s, '<'),
+		.pipes = str_cchr(s, '|'),
+	};
+}
