+++ test/lexer/test_pipes.c
@@ -1,6 +1,9 @@
-#include "../src/lexer.c"
+#include "../src/lexer/lexer.c"
+#include "../src/lexer/check_quotes.c"
+#include "../src/lexer/check_pipes.c"
+#include "../src/lexer/lexer_support.c"
 #include "../src/while_string_wrapper.c"
-#include "minishell.h"
+#include "../src/bool_array.c"
 #include "struct.h"
 #include "unity.h"
 #include <stdbool.h>
