+++ test/support/support_bool.h
@@ -15,7 +15,6 @@ t_test *returns_test_struct(const char *s, const char *expected);
 void	cleanup_test_struct(struct s_test *test);
 void	support_print(const char *s, const bool *arr);
 
-#include "unity.h"
-#include <string.h>
-#include <stdlib.h>
+void	range_ignore(const char *s, bool **ignore, unsigned char c);
+void	bool_arr_zeroing(const char *s, bool **ignore, size_t len);
 #endif
